+++ src/components/Modal/AddForm.js
@@ -1,5 +1,4 @@
 import React from "react";
-import classNames from "classnames";
 import {
     Button, Alert, FormGroup, Form, Input, Row, Col,
 } from "reactstrap";
