+++ src/routes.js
@@ -2,12 +2,12 @@ import Index from "views/Dashboard";
 import Placement from 'views/examples/Placement'
 import Internship from 'views/examples/Internship'
 import Profile from "views/examples/Profile.js";
-import Maps from "views/examples/Maps.js";
 import Register from "views/examples/Register.js";
 import SetPass from "views/examples/setPassword";
 import Login from "views/examples/Login.js";
-import Tables from "views/examples/Tables.js";
-import Icons from "views/examples/Icons.js";
+
+// import AdminInternship from 'views/AdminComponent/Internship'
+// import AdminPlacement from 'views/AdminComponent/Placement'
 
 var routes = [
   {
@@ -31,6 +31,20 @@ var routes = [
     component: Internship,
     layout: "/admin",
   },
+  // {
+  //   path: "/admin/placement",
+  //   name: "Admin Placement",
+  //   icon: "ni ni-tv-2 text-primary",
+  //   component: AdminPlacement,
+  //   layout: "/admin",
+  // },
+  // {
+  //   path: "/admin/internship",
+  //   name: "Admin Internship",
+  //   icon: "ni ni-tv-2 text-primary",
+  //   component: AdminInternship,
+  //   layout: "/admin",
+  // },
   {
     path: "/login",
     name: "Login",
@@ -46,8 +60,8 @@ var routes = [
     layout: "/auth",
   },
   {
-    path :"/user/register/verify/:code",
-    name :"SetPass",
+    path: "/user/register/verify/:code",
+    name: "SetPass",
     icon: "ni ni-circle-08 text-pink",
     component: SetPass,
     layout: "/auth",
