+++ src/views/AdminComponent/AdminTableContainer/InternshipTable.js
@@ -8,11 +8,15 @@ import {
     Button,
     Card, CardBody,
     Table, Row, Col,
-    Input, CustomInput,
+    CustomInput,
     Modal, ModalBody,
 } from "reactstrap";
 
 import Apply from '../../../components/Modal/ApplyForm'
+import Add from '../../../components/Modal/AddForm'
+
+import { ReactComponent as DownloadIcon } from '../../../assets/img/icons/common/save_alt_white_24dp.svg'
+import { ReactComponent as DeleteIcon } from '../../../assets/img/icons/common/delete_white_24dp.svg'
 
 import { tablestyle, applybtnshadow, headingstyle, expandbgstyle } from '../../../components/Style/css_style'
 
@@ -24,7 +28,10 @@ const TableContainer = ({ columns, data, }) => {
 
     const [fetchedData, setFetchedData] = useState([])
     const [isModal, setIsModal] = useState(false)
+    // Add offer Modal
     const [isAddModal, setIsAddModal] = useState(false)
+    // Delete Row confirmation Modal
+    const [isDeleteModal, setIsDeleteModal] = useState({ status: false, company: null })
 
     // For sorting columns
     const generateSortingIndicator = column => {
@@ -79,14 +86,57 @@ const TableContainer = ({ columns, data, }) => {
         setIsAddModal(!isAddModal)
     }
 
+    // After clicking on Download Icon
+    const handleDownload = () => {
+        alert("Downloaded")
+    }
+
+    // to handle delete modal confirmation 
+    const handleDeleteConfirm = () => {
+        // send the data to server to delete the data
+        alert("Confirmed")
+        // to remove Confirmation Modal after Confirming
+        handleDeleteModal()
+    }
+
+    // Modal to delete a row
+    const handleDeleteModal = (props) => {
+        if (isDeleteModal['status'] == true)
+            setIsDeleteModal({ status: false, company: null })
+        else
+            setIsDeleteModal({ status: !isDeleteModal['status'], company: props['cells'][1]['value'] })
+    }
+
 
     return (
         <Fragment>
+            {/*  */}
+            <Modal isOpen={isDeleteModal['status']} toggle={handleDeleteModal}>
+                <ModalBody>
+                    <hr />
+                    <h5>
+                        Are you sure, you want to delete the data of <b>{isDeleteModal['company']}</b>
+                    </h5>
+                    <hr />
+                    <div className='text-right mr-2'>
+                        <Button color='success' onClick={handleDeleteConfirm}>Confirm</Button>
+                        <Button outline onClick={handleDeleteModal}>Cancel</Button>
+                    </div>
+                    <hr />
+                </ModalBody>
+            </Modal>
+            {/* Modal for Organization specific form by click on Add Offer*/}
+            <Modal isOpen={isAddModal} toggle={handleModal}>
+                <ModalBody>
+                    <Add toggle={handleModal} />
+                </ModalBody>
+            </Modal>
             <Table responsive hover {...getTableProps()} style={tablestyle}>
                 <thead style={{ background: '#ccc' }}>
                     {headerGroups.map(headerGroup => (
                         <tr {...headerGroup.getHeaderGroupProps()}>
                             {headerGroup.headers.map(column => (
+                                // <th {...column.getHeaderProps()}>{column.render("Header")}</th>
                                 <th {...column.getHeaderProps(column.getSortByToggleProps())} style={{ textAlign: 'center' }}>
                                     <strong >{column.render("Header")}</strong>
                                     {generateSortingIndicator(column)}
@@ -96,19 +146,51 @@ const TableContainer = ({ columns, data, }) => {
                     ))}
                 </thead>
                 <tbody {...getTableBodyProps()}>
+                    <tr>
+                        <td colSpan='9' style={{ textAlign: 'center' }}>
+                            <Button onClick={handleModal} color='success' style={{ width: '100%' }}>
+                                Add Offer
+                            </Button>
+                        </td>
+                    </tr>
                     {page.map(row => {
                         prepareRow(row)
                         return (
                             <Fragment key={row.getRowProps().key}>
                                 <tr>
                                     {row.cells.map((cell, index) => {
+                                        let len = row.cells.length
                                         return <>
                                             {
-                                                // Starting 0-6 cells of a row
-                                                (index < 6) ?
+                                                // Starting cells of a row
+                                                (index < len - 2) ?
                                                     <td key={index} style={{ textAlign: 'center' }} {...cell.getCellProps()}>
                                                         {cell.render("Cell")}
                                                     </td> : <></>
+                                            }
+                                            {
+                                                //  Last Second cell of a row (Download Button)
+                                                (index === len - 2)
+                                                    ?
+                                                    <td key={index} className='text-center'>
+                                                        <Button color="success" size="sm" style={{ padding: '3px' }} onClick={handleDownload}>
+                                                            <DownloadIcon />
+                                                        </Button>
+                                                    </td>
+                                                    :
+                                                    <></>
+                                            }
+                                            {
+                                                // Last cell of a row (Delete Button)
+                                                (index === len - 1)
+                                                    ?
+                                                    <td key={index} className='text-center'>
+                                                        <Button color="danger" size="sm" style={{ padding: '3px' }} onClick={() => { handleDeleteModal(row) }}>
+                                                            <DeleteIcon />
+                                                        </Button>
+                                                    </td>
+                                                    :
+                                                    <></>
                                             }
                                         </>
                                     })}
@@ -116,7 +198,7 @@ const TableContainer = ({ columns, data, }) => {
                                 </tr>
                                 {row.isExpanded && (
                                     <tr>
-                                        <td colSpan='8'>{renderRowSubComponent(fetchedData, row.cells)}</td>
+                                        <td colSpan='9'>{renderRowSubComponent(fetchedData, row.cells)}</td>
                                     </tr>
                                 )}
                             </Fragment>
