+++ src/views/examples/TableContainer/PlacementTable.js
@@ -5,20 +5,16 @@ import { useTable, useSortBy, useExpanded, usePagination } from 'react-table'
 import "bootstrap/dist/css/bootstrap.min.css"
 // reactstrap component
 import {
-    Button, Container,
-    Card, CardHeader, CardBody,
-    NavItem, NavLink, Nav, Progress,
+    Button,
+    Card, CardBody,
     Table, Row, Col,
     Input, CustomInput,
-    Modal, ModalHeader, ModalBody, ModalFooter,
-    Form, FormGroup, Label, FormText
+    Modal, ModalBody,
 } from "reactstrap";
 
 import Apply from '../../../components/Modal/ApplyForm'
 import Add from '../../../components/Modal/AddForm'
 
-import { ReactComponent as DownloadIcon } from '../../../assets/img/icons/common/save_alt_white_24dp.svg'
-import { ReactComponent as DeleteIcon } from '../../../assets/img/icons/common/delete_white_24dp.svg'
 
 import { tablestyle, applybtnshadow, headingstyle, expandbgstyle } from '../../../components/Style/css_style'
 
@@ -31,7 +27,6 @@ const TableContainer = ({ columns, data, }) => {
     const [fetchedData, setFetchedData] = useState([])
     const [isModal, setIsModal] = useState(false)
     const [isAddModal, setIsAddModal] = useState(false)
-    const [isDeleteModal, setIsDeleteModal] = useState({ status: false, company: null })
 
     // For sorting columns
     const generateSortingIndicator = column => {
@@ -67,6 +62,7 @@ const TableContainer = ({ columns, data, }) => {
                         <Button style={applybtnshadow} color='success' onClick={toggle}>Apply</Button>
                     </CardBody>
                 </Card >
+                {/* Apply Form Modal */}
                 <Modal isOpen={isModal} toggle={toggle}>
                     <ModalBody>
                         <Apply data={cells} toggle={toggle} />
@@ -81,56 +77,9 @@ const TableContainer = ({ columns, data, }) => {
         setIsModal(!isModal)
     }
 
-    // Add Offer Modal
-    const handleModal = () => {
-        setIsAddModal(!isAddModal)
-    }
-
-    // After clicking on Download Icon
-    const handleDownload = () => {
-        alert("Downloaded")
-    }
-
-    // to handle delete modal confirmation 
-    const handleDeleteConfirm = () => {
-        // send the data to server to delete the data
-        alert("Confirmed")
-        // to remove Confirmation Modal after Confirming
-        handleDeleteModal()
-    }
-
-    // Modal to delete a row
-    const handleDeleteModal = (props) => {
-        if (isDeleteModal['status'] == true)
-            setIsDeleteModal({ status: false, company: null })
-        else
-            setIsDeleteModal({ status: !isDeleteModal['status'], company: props['cells'][1]['value'] })
-    }
-
 
     return (
         <Fragment>
-            {/*  */}
-            <Modal isOpen={isDeleteModal['status']} toggle={handleDeleteModal}>
-                <ModalBody>
-                    <hr />
-                    <h5>
-                        Are you sure, you want to delete the data of <b>{isDeleteModal['company']}</b>
-                    </h5>
-                    <hr />
-                    <div className='text-right mr-2'>
-                        <Button color='success' onClick={handleDeleteConfirm}>Confirm</Button>
-                        <Button outline onClick={handleDeleteModal}>Cancel</Button>
-                    </div>
-                    <hr />
-                </ModalBody>
-            </Modal>
-            {/* Modal for Organization specific form by click on Apply*/}
-            <Modal isOpen={isAddModal} toggle={handleModal}>
-                <ModalBody>
-                    <Add toggle={handleModal} />
-                </ModalBody>
-            </Modal>
             <Table responsive hover {...getTableProps()} style={tablestyle}>
                 <thead style={{ background: '#ccc' }}>
                     {headerGroups.map(headerGroup => (
@@ -145,13 +94,6 @@ const TableContainer = ({ columns, data, }) => {
                         </tr>
                     ))}
                 </thead>
-                <tr>
-                    <td colSpan='8' style={{ textAlign: 'center' }}>
-                        <Button onClick={handleModal} color='success' style={{ width: '100%' }}>
-                            Add Offer
-                        </Button>
-                    </td>
-                </tr>
                 <tbody {...getTableBodyProps()}>
                     {page.map(row => {
                         prepareRow(row)
@@ -166,30 +108,6 @@ const TableContainer = ({ columns, data, }) => {
                                                     <td key={index} style={{ textAlign: 'center' }} {...cell.getCellProps()}>
                                                         {cell.render("Cell")}
                                                     </td> : <></>
-                                            }
-                                            {
-                                                //  7th cell of a row (Download Button)
-                                                (index == 6)
-                                                    ?
-                                                    <td key={index} className='text-center'>
-                                                        <Button color="success" size="sm" style={{ padding: '3px' }} onClick={handleDownload}>
-                                                            <DownloadIcon />
-                                                        </Button>
-                                                    </td>
-                                                    :
-                                                    <></>
-                                            }
-                                            {
-                                                // 8th cell of a row (Delete Button)
-                                                (index == 7)
-                                                    ?
-                                                    <td key={index} className='text-center'>
-                                                        <Button color="danger" size="sm" style={{ padding: '3px' }} onClick={() => { handleDeleteModal(row) }}>
-                                                            <DeleteIcon />
-                                                        </Button>
-                                                    </td>
-                                                    :
-                                                    <></>
                                             }
                                         </>
                                     })}
